requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.5.30
cloudinary==1.26.0
config==0.5.0.post0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==4.0.3
django-bootstrap3==15.0.0
django-bootstrap4==3.0.1
django-heroku==0.3.1
django-registration==2.4
django-utils-six==2.0
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.3.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.2.0
